Move test files into packages
