+++ src/test/java/seedu/bankwithus/storage/StorageTest.java
@@ -1,0 +1,14 @@
+//@@author tyuyang
+package seedu.bankwithus.storage;
+
+import org.junit.jupiter.api.Test;
+import seedu.bankwithus.storage.Storage;
+
+import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
+
+public class StorageTest {
+    @Test
+    public void load_normalSaveFile_noMessage() {
+        assertDoesNotThrow(() -> new Storage("data/save.txt", "data/transaction.txt"));
+    }
+}
